AEA-1507: Add ODS API check to the _status endpoint. (#582)
* AEA-1507: Add ODS API check to the _status endpoint.

* AEA-1507: Add hapi-pino types to pact project.

* AEA-1507: Separate exported status check code from the handler code to avoid horrible TypeScript compilation issues.
