+++ coordinator/src/routes/health/get-status.ts
@@ -1,42 +1,14 @@
 import Hapi from "@hapi/hapi"
-import axios, {AxiosError} from "axios"
+import axios from "axios"
 import {VALIDATOR_HOST} from "../util"
 import {spineClient} from "../../services/communication/spine-client"
-import pino from "pino"
-
-export interface StatusCheckResponse {
-  status: "pass" | "warn" | "error"
-  timeout: "true" | "false"
-  responseCode: number
-  outcome?: string
-  links?: string
-}
-
-export async function serviceHealthCheck(url: string, logger: pino.Logger): Promise<StatusCheckResponse> {
-  try {
-    const response = await axios.get<string>(url, {timeout: 20000})
-    return {
-      status: response.status === 200 ? "pass" : "error",
-      timeout: "false",
-      responseCode: response.status,
-      outcome: response.data,
-      links: url
-    }
-  } catch (error) {
-    logger.error("Error calling external service for status check: " + error.message)
-    const axiosError = error as AxiosError
-    return {
-      status: "error",
-      timeout: axiosError.code === "ECONNABORTED" ? "true" : "false",
-      responseCode: axiosError.response?.status,
-      outcome: axiosError.response?.data,
-      links: url
-    }
-  }
-}
+import {odsClient} from "../../services/communication/ods-client"
+import {serviceHealthCheck, StatusCheckResponse} from "../../utils/status"
 
 function createStatusResponse(
-  errorStatusCode: number, checks: Record<string, Array<StatusCheckResponse>>, h: Hapi.ResponseToolkit
+  errorStatusCode: number,
+  checks: Record<string, Array<StatusCheckResponse>>,
+  h: Hapi.ResponseToolkit
 ) {
   let responseStatus = "pass"
   let responseCode = 200
@@ -64,6 +36,7 @@ export default [
     handler: async (request: Hapi.Request, h: Hapi.ResponseToolkit): Promise<Hapi.ResponseObject> => {
       return createStatusResponse(200, {
         "validator:status": [await serviceHealthCheck(`${VALIDATOR_HOST}/_status`, request.logger)],
+        "ods:status": [await odsClient.getStatus(request.logger)],
         "spine:status": [await spineClient.getStatus(request.logger)]
       }, h)
     }
