+++ coordinator/src/services/communication/spine-client.ts
@@ -1,12 +1,12 @@
 import {LiveSpineClient} from "./live-spine-client"
 import {SandboxSpineClient} from "./sandbox-spine-client"
 import {spine} from "@models"
-import pino, {Logger} from "pino"
-import {StatusCheckResponse} from "../../routes/health/get-status"
+import pino from "pino"
+import {StatusCheckResponse} from "../../utils/status"
 
 export interface SpineClient {
-  send(spineRequest: spine.SpineRequest, logger: Logger): Promise<spine.SpineResponse<unknown>>
-  poll(path: string, fromAsid: string, logger: Logger): Promise<spine.SpineResponse<unknown>>
+  send(spineRequest: spine.SpineRequest, logger: pino.Logger): Promise<spine.SpineResponse<unknown>>
+  poll(path: string, fromAsid: string, logger: pino.Logger): Promise<spine.SpineResponse<unknown>>
   getStatus(logger: pino.Logger): Promise<StatusCheckResponse>
 }
