+++ tests/e2e/pact/package.json
@@ -26,6 +26,7 @@
     "@pact-foundation/pact": "^10.0.0-beta.34",
     "@types/axios": "^0.14.0",
     "@types/crypto-js": "^4.0.1",
+    "@types/hapi-pino": "^8.0.3",
     "@types/jest": "24.9.0",
     "@types/jsonwebtoken": "^8.5.0",
     "@types/lossless-json": "1.0.0",
